requirements.txt
flask
flask_login
oauthlib
pyopenssl
requests
azure_cosmos